update stream read SplitAssigner
