+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/sink/RowHashKeyChannelComputer.java
@@ -18,6 +18,7 @@
 
 package org.apache.paimon.flink.sink;
 
+import org.apache.paimon.flink.ChannelComputer;
 import org.apache.paimon.schema.TableSchema;
 
 import org.apache.flink.table.data.RowData;
