+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/source/operator/MonitorFunction.java
@@ -18,6 +18,8 @@
 
 package org.apache.paimon.flink.source.operator;
 
+import org.apache.flink.api.common.functions.Partitioner;
+import org.apache.paimon.flink.ChannelComputer;
 import org.apache.paimon.flink.utils.JavaTypeInfo;
 import org.apache.paimon.table.source.DataSplit;
 import org.apache.paimon.table.source.EndOfScanException;
@@ -235,9 +237,10 @@ public class MonitorFunction extends RichSourceFunction<Split>
                         name + "-Monitor",
                         new JavaTypeInfo<>(Split.class))
                 .forceNonParallel()
-                .partitionCustom(
-                        (key, numPartitions) -> key % numPartitions,
-                        split -> ((DataSplit) split).bucket())
+                .partitionCustom((Partitioner<Split>) (split, numPartitions) ->
+                        ChannelComputer.select(
+                                ((DataSplit) split).partition(),
+                                ((DataSplit) split).bucket(), numPartitions), split -> split)
                 .transform(name + "-Reader", typeInfo, new ReadOperator(readBuilder));
     }
 }
